refactor: merge packages
