--- packages/mcp-manager/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@kortex-hub/mcp-manager",
-  "version": "0.2.0-next",
-  "description": "",
-  "main": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "license": "Apache-2.0",
-  "repository": "https://github.com/kortex-hub/mcp-runner",
-  "publishConfig": {
-    "provenance": true,
-    "access": "public"
-  },
-  "scripts": {
-    "build": "vite build",
-    "test": "vitest run --coverage --passWithNoTests"
-  },
-  "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.18.2"
-  },
-  "devDependencies": {
-    "typescript": "5.9.3",
-    "vite": "^7.1.3",
-    "vitest": "^3.0.9",
-    "vite-plugin-dts": "^4.5.4",
-    "@kortex-hub/mcp-registry-types": "^1.2.3",
-    "@kortex-hub/mcp-runner": "workspace:*"
-  },
-  "files": [
-    "dist"
-  ]
-}
