--- packages/mcp-manager/vite.config.ts
@@ -1,63 +1,0 @@
-/**********************************************************************
- * Copyright (C) 2025 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ***********************************************************************/
-/* eslint-env node */
-import { join } from 'node:path';
-import { defineConfig } from 'vite';
-import { builtinModules } from 'node:module';
-import dts from 'vite-plugin-dts';
-
-const PACKAGE_ROOT = __dirname;
-
-// https://vitejs.dev/config/
-export default defineConfig({
-    mode: process.env.MODE,
-    root: PACKAGE_ROOT,
-    resolve: {
-        alias: {
-            '/@/': join(PACKAGE_ROOT, 'src') + '/',
-        },
-    },
-    plugins: [dts({
-        exclude: ['tests/**/*', 'src/**/*.{test,spec}.ts']
-    })],
-    build: {
-        target: 'esnext',
-        sourcemap: 'inline',
-        outDir: 'dist',
-        assetsDir: '.',
-        minify: process.env.MODE === 'production' ? 'esbuild' : false,
-        lib: {
-            fileName: 'index',
-            entry: 'src/index.ts',
-            formats: ['es'],
-        },
-        rollupOptions: {
-            external: builtinModules.flatMap(p => [p, `node:${p}`]),
-        },
-        emptyOutDir: true,
-        reportCompressedSize: false,
-    },
-    test: {
-        environment: 'node',
-        include: [
-            'src/**/*.{test,spec}.?(c|m)[jt]s?(x)',
-            'tests/**/*.{test,spec}.?(c|m)[jt]s?(x)',
-        ],
-        passWithNoTests: true,
-    },
-});
