--- packages/mcp-registry-client/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "@tsconfig/node22/tsconfig.json",
-  "include": ["src/**/*.ts"],
-}
