+++ packages/mcp-runner/src/manager/mcp-manager.ts
@@ -19,8 +19,8 @@ import {Storage} from "/@/models/storage";
 import {MCPInstance} from "/@/models/mcp-instance";
 import type { components } from '@kortex-hub/mcp-registry-types';
 import { randomUUID } from 'node:crypto';
-import { MCPRemote, MCPPackage } from '@kortex-hub/mcp-runner';
-import type  { MCPRemoteOptions } from "@kortex-hub/mcp-runner";
+import {MCPRemote, MCPRemoteOptions} from "/@/remote/mcp-remote.js";
+import {MCPPackage} from "/@/package/mcp-package.js";
 import {RemoteConfig} from "/@/models/remote-config";
 import {PackageConfig} from "/@/models/package-config";
 import { EventEmitter } from 'node:events';
