+++ packages/mcp-runner/src/models/package-config.ts
