+++ packages/mcp-runner/src/utils/format-key-value-inputs.ts
