+++ packages/mcp-runner/src/utils/input-with-variables.ts
